refactor clients to util and use flow names not IDs in mcp.py
