+++ src/backend/base/langflow/base/mcp/util.py
@@ -1,10 +1,18 @@
 import asyncio
+import os
 from collections.abc import Awaitable, Callable
+from contextlib import AsyncExitStack
 from typing import Any
+from uuid import UUID
 
+import httpx
+from mcp import ClientSession, StdioServerParameters, stdio_client
+from mcp.client.sse import sse_client
 from pydantic import Field, create_model
+from sqlmodel import select
 
 from langflow.helpers.base_model import BaseModel
+from langflow.services.database.models import Flow
 
 
 def create_tool_coroutine(tool_name: str, arg_schema: type[BaseModel], session) -> Callable[..., Awaitable]:
@@ -52,6 +60,17 @@ def create_tool_func(tool_name: str, arg_schema: type[BaseModel], session) -> Ca
     return tool_func
 
 
+async def get_flow(flow_name: str, user_id: str, session) -> Flow | None:
+    uuid_user_id = UUID(user_id) if isinstance(user_id, str) else user_id
+    stmt = select(Flow).where(Flow.user_id == uuid_user_id).where(Flow.is_component == False)  # noqa: E712
+    flows = (await session.exec(stmt)).all()
+
+    for flow in flows:
+        if flow.to_data().name == flow_name:
+            return flow
+    return None
+
+
 def create_input_schema_from_json_schema(schema: dict[str, Any]) -> type[BaseModel]:
     """Converts a JSON schema into a Pydantic model dynamically.
 
@@ -91,3 +110,72 @@ def create_input_schema_from_json_schema(schema: dict[str, Any]) -> type[BaseMod
         fields[field_name] = (base_type, Field(**field_metadata))
 
     return create_model("InputSchema", **fields)
+
+
+class MCPStdioClient:
+    def __init__(self):
+        self.session: ClientSession | None = None
+        self.exit_stack = AsyncExitStack()
+
+    async def connect_to_server(self, command_str: str):
+        command = command_str.split(" ")
+        server_params = StdioServerParameters(
+            command=command[0],
+            args=command[1:],
+            env={"DEBUG": "true", "PATH": os.environ["PATH"]},
+        )
+        stdio_transport = await self.exit_stack.enter_async_context(stdio_client(server_params))
+        self.stdio, self.write = stdio_transport
+        self.session = await self.exit_stack.enter_async_context(ClientSession(self.stdio, self.write))
+        await self.session.initialize()
+        response = await self.session.list_tools()
+        return response.tools
+
+
+# Define constant for status code
+HTTP_TEMPORARY_REDIRECT = 307
+
+
+class MCPSseClient:
+    def __init__(self):
+        self.write = None
+        self.sse = None
+        self.session: ClientSession | None = None
+        self.exit_stack = AsyncExitStack()
+
+    async def pre_check_redirect(self, url: str):
+        async with httpx.AsyncClient(follow_redirects=False) as client:
+            response = await client.request("HEAD", url)
+            if response.status_code == HTTP_TEMPORARY_REDIRECT:
+                return response.headers.get("Location")
+        return url
+
+    async def _connect_with_timeout(
+        self, url: str, headers: dict[str, str] | None, timeout_seconds: int, sse_read_timeout_seconds: int
+    ):
+        sse_transport = await self.exit_stack.enter_async_context(
+            sse_client(url, headers, timeout_seconds, sse_read_timeout_seconds)
+        )
+        self.sse, self.write = sse_transport
+        self.session = await self.exit_stack.enter_async_context(ClientSession(self.sse, self.write))
+        await self.session.initialize()
+
+    async def connect_to_server(
+        self, url: str, headers: dict[str, str] | None, timeout_seconds: int = 500, sse_read_timeout_seconds: int = 500
+    ):
+        if headers is None:
+            headers = {}
+        url = await self.pre_check_redirect(url)
+        try:
+            await asyncio.wait_for(
+                self._connect_with_timeout(url, headers, timeout_seconds, sse_read_timeout_seconds),
+                timeout=timeout_seconds,
+            )
+            if self.session is None:
+                msg = "Session not initialized"
+                raise ValueError(msg)
+            response = await self.session.list_tools()
+        except asyncio.TimeoutError as err:
+            msg = f"Connection to {url} timed out after {timeout_seconds} seconds"
+            raise TimeoutError(msg) from err
+        return response.tools
