+++ src/backend/base/langflow/components/deactivated/mcp_sse.py
@@ -1,68 +1,17 @@
 # from langflow.field_typing import Data
-import asyncio
-from contextlib import AsyncExitStack
 
-import httpx
 from langchain_core.tools import StructuredTool
-from mcp import ClientSession, types
-from mcp.client.sse import sse_client
+from mcp import types
 
-from langflow.base.mcp.util import create_input_schema_from_json_schema, create_tool_coroutine, create_tool_func
+from langflow.base.mcp.util import (
+    MCPSseClient,
+    create_input_schema_from_json_schema,
+    create_tool_coroutine,
+    create_tool_func,
+)
 from langflow.custom import Component
 from langflow.field_typing import Tool
 from langflow.io import MessageTextInput, Output
-
-# Define constant for status code
-HTTP_TEMPORARY_REDIRECT = 307
-
-
-class MCPSseClient:
-    def __init__(self):
-        # Initialize session and client objects
-        self.write = None
-        self.sse = None
-        self.session: ClientSession | None = None
-        self.exit_stack = AsyncExitStack()
-
-    async def pre_check_redirect(self, url: str):
-        """Check if the URL responds with a 307 Redirect."""
-        async with httpx.AsyncClient(follow_redirects=False) as client:
-            response = await client.request("HEAD", url)
-            if response.status_code == HTTP_TEMPORARY_REDIRECT:
-                return response.headers.get("Location")  # Return the redirect URL
-        return url  # Return the original URL if no redirect
-
-    async def _connect_with_timeout(
-        self, url: str, headers: dict[str, str] | None, timeout_seconds: int, sse_read_timeout_seconds: int
-    ):
-        """Connect to the SSE server with timeout."""
-        sse_transport = await self.exit_stack.enter_async_context(
-            sse_client(url, headers, timeout_seconds, sse_read_timeout_seconds)
-        )
-        self.sse, self.write = sse_transport
-        self.session = await self.exit_stack.enter_async_context(ClientSession(self.sse, self.write))
-        await self.session.initialize()
-
-    async def connect_to_server(
-        self, url: str, headers: dict[str, str] | None, timeout_seconds: int = 500, sse_read_timeout_seconds: int = 500
-    ):
-        if headers is None:
-            headers = {}
-        url = await self.pre_check_redirect(url)
-        try:
-            await asyncio.wait_for(
-                self._connect_with_timeout(url, headers, timeout_seconds, sse_read_timeout_seconds),
-                timeout=timeout_seconds,
-            )
-            # List available tools
-            if self.session is None:
-                msg = "Session not initialized"
-                raise ValueError(msg)
-            response = await self.session.list_tools()
-        except asyncio.TimeoutError as err:
-            error_message = f"Connection to {url} timed out after {timeout_seconds} seconds"
-            raise TimeoutError(error_message) from err
-        return response.tools
 
 
 class MCPSse(Component):
